[IMP] loyalty_partner_applicability: Allow to share coupons inside same entity
There's a new configuration option in the settings to allow sharing
coupons between members of the same trading entity.

It has to be in the base module for being shared between all the rest of
applications, but each *_applicability module will contain extra bits of
code to use it.

TT45413