[IMP] loyalty_mass_mailing: Adapt the module to the new way of applying loyalty programs to partners
The partner applicability of a loyalty rule can now be defined at the level of the loyalty program
in addition to the level of the loyalty rule. This commit adapts the module to this new way of
defining the partner applicability of a loyalty program when computing the domain to use for
the mass mailing of a loyalty program.